+++ src/plugins/vis_type_vega/public/data_model/vega_parser.test.js
@@ -33,6 +33,7 @@ import { euiThemeVars } from '@osd/ui-shared-deps/theme';
 import { euiPaletteColorBlind } from '@elastic/eui';
 import { VegaParser } from './vega_parser';
 import { bypassExternalUrlCheck } from '../vega_view/vega_base_view';
+import { VisLayerTypes } from '../../../vis_augmenter/public';
 
 jest.mock('../services');
 
@@ -388,6 +389,25 @@ describe('VegaParser._parseConfig', () => {
     check({ config: { kibana: { a: 1 } } }, { a: 1 }, { config: {} })
   );
   test('_hostConfig', check({ _hostConfig: { a: 1 } }, { a: 1 }, {}, 1));
+  test(
+    'visibleVisLayers conversion to map',
+    check(
+      {
+        config: {
+          kibana: {
+            hideWarnings: true,
+            visibleVisLayers: [[VisLayerTypes.PointInTimeEvents, true]],
+          },
+        },
+      },
+      {
+        hideWarnings: true,
+        visibleVisLayers: new Map([[VisLayerTypes.PointInTimeEvents, true]]),
+      },
+      { config: {} },
+      0
+    )
+  );
 });
 
 describe('VegaParser._compileWithAutosize', () => {